# src/main/java/com/example/computershop/config/ErrorMessageResponse.java
package com.example.computershop.config;

import com.fasterxml.jackson.annotation.JsonFormat;

import java.time.LocalDateTime;

public record ErrorMessageResponse (
        String message,
        String detailedMessage,
        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
        LocalDateTime dateTime
) {}